Make buck documentation publishing its own job
Summary:This adds a new shipit job for updating buck documetation.

It pulls two copies of the opensource repo, one to generate the docs from, and
one for the gh-pages branch where the docs will go.

Doc changes are committed regularly instead of always overwriting history, so
they can be properly versioned. No commits will be made if docs do not change.
The public github hash is included in the commit message so doc version can be
synced to code version if need be.

In the future we can also have the documentation generator exit early if no
commits have been made, whereas right now it will always run the generator.

Since this rule is not actually touching external repos, it would be
conceivable that we use travis ci for this. But I've heard we had some issues
with travis before.

Reviewed By: fredemmott

Differential Revision: D3095717

fb-gh-sync-id: 04b51fdc08138e75d65c8eb764bf31a12d321165
fbshipit-source-id: 04b51fdc08138e75d65c8eb764bf31a12d321165