Shipit for OpenBMC repo
Summary:Add OpenBMC to shipit.
There is a bit of work needed to support a branch with merge commits.
It is possible to script a git lbg that cherry-picks the commits into a new 'amalgamation branch'.
This branch, with --skip-init, and --skip-pull can be synced to the github repo.

As follow ups we will need to land the kernel, u-boot, and lightning code before running shipit for OpenBMC in chronos.

Reviewed By: fredemmott

Differential Revision: D3138172

fb-gh-sync-id: e7364ac5bf2497154e6be171184d57eec5521123
fbshipit-source-id: e7364ac5bf2497154e6be171184d57eec5521123