Add Github overlay for OpenBMC
Summary:See the changes in D3170312.
This overlays the directory structure within `./github/` onto `./`.

Before this can be used the destination repository must be reverted or pruned of the overlay files.
The Github project, within the opensource tool, must also reflect the overlay directories in order to accept public contributions.
This is for the most part simple if the overlays are restricted to u-boot and the Linux kernel.

Reviewed By: jmehta

Differential Revision: D3170412

fb-gh-sync-id: 49bb14ea80954244b14a338b813d5954350d7ae7
fbshipit-source-id: 49bb14ea80954244b14a338b813d5954350d7ae7