set git user for new repos
Summary:
Fixes `shipit_fbshipit.php` and `--create-new-repo` for all projects
when running as a user with no git config, and no unix user info.

Fake users are only used when the commit is never actually exposed; for the real
case (`--create-new-repo` for 'initial commit'), the normal github bot credentials are used.

Reviewed By: jamesgpearce

Differential Revision: D3244128

fbshipit-source-id: fd2deaf0f5f2647ff13fe234d9b30d9c990221a7