Allow overriding committer name/email
Summary:
adds `--destination-committer-name` and `--destination-committer-email` options which can be used to set/override the name and email of the committer.

also modifies `ShipItCreateNewRepoPhase` so that you can pass a callback that determines the committer info for the new repo's initial commit. for example:

```
$config ==> shape(
  'name' => $config->getDestinationCommitterName() ?? "Herp",
  'email' => $config->getDestinationCommitterEmail() ?? "Derp",
)
```

(in reality you'd probably want to fall back to asking `ShipItGitHubUtils::getCredentialsForProject` if not set in the config)

Reviewed By: sdwilsh

Differential Revision: D3357798

fbshipit-source-id: 0fcbcd070647df347c0a93333a69f785037f69c1