add github pr sync to list of supported sections
Summary:
Buck shipit jobs was not properly attributing the author of a pull request,
instead opting for the committer.

It seems that in such cases the message contains a `Github Author:` line
followed by `Github PR Sync` line. As the latter was not a supported flag,
`ShipItMessageSections::getSections` parsed the github author as a multi-line
field, and then when re-emitting it, added a leading newline.

This only affects buck because persumably only buck does custom message
rewriting before passing the message to FBCommonFilters.

Reviewed By: fredemmott

Differential Revision: D3401916

fbshipit-source-id: cb82158eedaa020d630ad4a7874ffaf85820d9e4