remove whitelist
Summary:
Depends on D3450250

D3450250 (in the Infer repo) is the top of a stack of diffs that move files
around to end up with a clearer situation of which files are and are not
blacklisted from GitHub in the internal Infer repo.

This diff takes advantage of that to remove the blacklist/whitelist mechanism
in Infer in favour of a simpler blacklist-only one.

Here's how I came up with the list of things to blacklist and to test:

- Get the list of toplevel files, stare at them and copy only the ones we want blacklisted:

  git ls-files | grep -e '^[^/]\+$'

- Get the list of toplevel directories, stare at them and copy only the ones we want blacklisted:

  git ls-files | grep -e '^[^/]\+/' | sed -e "s#^\([^/]*\)/.*\$#      '\1/',#g" | sort | uniq

- run the tests `./fb_runtests.sh`. There were many failures. For each of them
  complaining that file or directory `<path>` was not blacklisted, double check
  that the path does not exist in Infer anymore with

  git ls-files <path>

  then remove `<path>` from the tests.

Reviewed By: fredemmott

Differential Revision: D3457371

fbshipit-source-id: fec220d225bc7c4621b65202cfbfce25f655b52a