catch repo corruption when it happens, retry
Summary:
Previous fixes would detect corrupted repos and throw them away - but,
the pull that created the corruption would still fail, as would the job as a
whole. This was intentional to get some debugging data, and because it didn't
happen that often. Now it's common, we need to work around it.

Reviewed By: sdwilsh

Differential Revision: D3531780

fbshipit-source-id: 0a84b59ffc4870eedb9197c65ea18618e241128b