Fixing Shipit script shipping TARGETS file
Summary:
There was a change to ShipIt a while back that picked up the TARGETS file during stripCommonFiles/exceptions. We need to add "@/fb/@" because otherwise the entire path would get stripped, I changed that to make it more clear.

I'm also removing the /facebook/ directory since that's no longer needed with ShipIt.

Reviewed By: JoelMarcey

Differential Revision: D3831186

fbshipit-source-id: 20e4233b14a9121076f1330148f4186014af7904