Prepare shipit and tagit for css-layout -> Yoga
Summary:
We are renaming `css-layout` to `Yoga`. This changes the `shipit` and `tagit` code to reflect that change.

I will land this the morning of the launch, and then push the new `shipit` and `tagit` packages just after launch.

Note: The generation of the autoload map for `tagit` brought in some other changes that were, I guess, not run in other commits.

Reviewed By: fredemmott

Differential Revision: D4277911

fbshipit-source-id: a71041abb6ca794b0e1bcbfa74dbf3edaa4b8e93