Update OpenBMC ShipIt tools to support fido and helium
Summary:
This updates OpenBMC's three ShipIt scripts that publish the open source OpenBMC.
The internal project is developed in a version 1 (fido) and 2 (helium) cadence with the released/mature platforms using fido.

That means we continue to receive updates to fido and must sync them publically. These land into the `master` branch on Github.
We have always wanted to publish the in-development version helium but the history was non-linear and needed lots of resolution love.
The root of the rOPENBMC repository in the `helium` branch contained several merge commits from its `fido` and `master` branch.

Those have been resolved and published to Github's repo under the `helium` branch.
Future commits to `helium` will sync fine, and we're now encouraging and only accepting cherry-picks from `fido` (or otherwise).

To make this work from an OSS perspective we are now publishing 6 branches.
This is because `helium`'s branch uses a newer version of U-Boot and Linux-- thus their scripts are updated.

Reviewed By: fredemmott

Differential Revision: D4376874

fbshipit-source-id: 776e84d7b071eaa16835c21f2524c79f93ef31f4