Use the proper config object
Summary:
When we update the config before we start running phases, a new config object
is created, which means our getter was not using the correct config, so paths
were incorrect.

The getter is needed because otherwise we won't stick with the revision we
update to, since every time we create a repo object, we try to update to
the appropriate branch.

Reviewed By: fredemmott

Differential Revision: D4510718

fbshipit-source-id: 9f84960f6b54af3f74e4a664b7c82220cdd4c4ee