Part 1 of eliminating most of the per-project boilerplate
Summary:
I'm pushing a lot more of the docs and logic that can be shared between projects into `fbcode_builder`.

On the ShipIt side, I created a filter that hides all the details of `fbcode_builder` setup from the project owner, so that adding the filter is now a 5-line diff.

Up next: eliminating the 3 remaining per-project scripts that I have in Bistro, and folding them into `fbcode_builder`.

Reviewed By: sdwilsh

Differential Revision: D4474207

fbshipit-source-id: ea27a537e46baad004e898736127fe52dc8056cd