Ensure that every ShipIt executable tests every path mapping
Summary:
This adds a test that ensures that all test files for ShipIt executables
actually test every path mapping they declare.

This actually caught a bug in one existing ShipIt executable, and it's
something we've been looking for by hand when we review these.  Machines can
do this faster and better.

Reviewed By: JoelMarcey

Differential Revision: D4519322

fbshipit-source-id: c8c125ba7a6025d71ee5a760668389ab49de152b