ImportIt files for react-native
Summary:
ImportIt is a new, easier way to import changes into Phabricator.  This will
replace the path mappings on the Advanced Settings page for your project.

This is a little more complicated than most projects because you rename some files
for flow, but existing tests caught it and I was able to easily make the needed
changes.

Reviewed By: zertosh

Differential Revision: D4552740

fbshipit-source-id: 3b6bf3f96ae8a43c21dfb61a03c31c9944d97e86