Properly handle merge commits with ImportIt
Summary:
If a pull request has merges from master back to the branch it is being
developed on, those commits do not end up in the patch file that GitHub
generates.  However, we cannot just use the diff file that GitHub generates
either because that does not contained changes to binary files.

This change has ImportIt first fetch the pull request changes (via
`refs/pull/{pr #}/head`), compute the merge base with the source branch (the
default branch ShipIt publishes to for the GitHub project), produce a binary
diff from the PR head revision and the merge-base, and finally commits that
change on top of the merge-base.

At this point, ImportIt can generate a changeset, and then apply it to the
destination repository.

Reviewed By: JoelMarcey

Differential Revision: D4629259

fbshipit-source-id: 511ded51a940c970537287f9d81771d1e0820b72