ship the fboss/common stubs in eden as well
Summary:
Ship the fboss/common stub files in the eden repository too.

This also adds a rule to rename externals/TARGETS.oss to externals/TARGETS.
I will add this file in an upcoming diff to get the eden opensource build
working with buck.  This TARGETS file contains the rules for dependencies that
are external to our github repository.  It is only applicable to the github
repository, and should not be used in our internal repository.  Therefore we
need to make sure it isn't named TARGETS in the Facebook-internal repository.
This uses the moveDirectories() filter for this rename, even though this isn't
really a directory.

Reviewed By: bolinfest

Differential Revision: D4633431

fbshipit-source-id: ac73ea1585d684964219f151558f0a537668b424