Provide better error messages when mapping tests fail
Summary:
The default array comparison message when path mapping tests fail is not
intuitive and hard to understand.  This changes the error message to display
what was expected to happen, and also include any debug messaging (which
includes things like what was stripped and why) which makes this *much* simpler
to debug.

Reviewed By: caabernathy, gfosco

Differential Revision: D4820788

fbshipit-source-id: d011c17967388978d8c04499b4978b79825f2c3d