Properly handle submodule additions and removals.
Summary:
This allows doing things like `--verify --create-fixup-patch` to correctly
update submodules, but also makes it easier for us to handle adding submodules
in the future for projects and ShipIt can automatically handle it.

This also adds an integration test to verify that adding, modifying, and
removing a submodule works properly with Git.

Reviewed By: fredemmott

Differential Revision: D4943173

fbshipit-source-id: 136669ac29c07863952ce2944feeea55918039ec