Hardcode explicit exception to default submodule path for non-fbsource repos
Summary:
Primarily the issue here is that we do several git operations over the root paths. This is a problem when the path doesn't exist (as is the case for many projects in standalone repos). Some of it is solvable (`git log --ancestry-path -- path1 path2` instead of `git log --ancestry-path path1 path2`), but not for `git archive`. So until we fix all of this in a better way, the quick fix is to just never add the bad root when we know it is impossible for it to exist.

This should be replaced with something better, perhaps making this path opt-in to begin with.

Reviewed By: sdwilsh

Differential Revision: D4966419

fbshipit-source-id: 9f595b946cf524bb130eafcf55292aa11853f8d7