Add support for OAuth token in addition to user/pass
Summary:
In `makeAPIRequest` we were only using our authentication for the initial request, but pagination didn't have the auth info, which we were prepending into the URL. Generally this is probably fine, but un-authed requests have a rate limit of 60 per hour, so it's entirely plausible that you could hit that.

There were 2 fixes that led naturally from this - the first is to actually set the HTTP basic auth via curl directly, instead of doing it in the URL. From there it was straightforward to just add support for OAuth tokens directly. They can be used interchangabley (token can be used as the password for basic auth here), but it's clearer to actually use token support.

Reviewed By: sdwilsh

Differential Revision: D5297482

fbshipit-source-id: e261f965e7babbae45a003fdf52bbb3410b28874