Add configs for Draft.js
Summary:
~what~
We are re-adding the 'ship-it' configs that were deleted in D4530046
I changed the root paths because they seem outdated; we don't have a
`draft-js/__github__/` directory internally, and we no longer use
`static_upstream`.[1]

[1]: https://fb.facebook.com/groups/jsinfra/permalink/1104942092887738/

~why~
Right now we need to run a Draft.js sync manually in order to pull in any
change from open source. This can take hours to days, and is prohibitively
costly, to the point where we have virtually stopped merging changes from OS.

We hope that enabling ship-it will make this step faster and easier, because
ideally we will run a sync with every change made to Draft.js in OS.

Lastly - right now Draft.js is out of sync between www and Github. It has not
been feasible to keep them in sync by hand - people internally at FB make
changes to Draft faster than we can review them. Running syncs more often will
help with this problem as well.

Reviewed By: spicyj

Differential Revision: D5339943

fbshipit-source-id: 86940b0b1f4de824127795d7d1a5372173190c00