Stop stripping .npmrc, .yarnrc and yarn.lock
Summary: Since yarn 0.23.0, offline deps use urls for the "resolved" paths, which makes the `yarn.lock` portable across environment with or without an offline mirror. So we can start syncing it now. The npmrc was previously kept from open source because it had the fwdproxy stuff in it - we don't use that anymore.

Reviewed By: zpao

Differential Revision: D5478696

fbshipit-source-id: 0a53e99546cd408625beac1c1a6c50b9b276881a