fix importit so that more specific path move rules take precedence
Summary:
Fix ImportItPathFilters so that more-specific path filters take precedence over
less-specific ones.

Previously the code simply used the reverse ordering of the shipit filters,
which is not necessarily correct.  I added a unit test demonstrating this,
which failed with the old code, and passes with the new code.

It doesn't seem possible to write an automatic version of
ImportItPathFilters::moveDirectories() that is correct in all cases, but this
new behavior certainly seems like it should be correct in more cases than the
old code.  All existing importit tests pass with the new behavior.

Reviewed By: sdwilsh

Differential Revision: D4633433

fbshipit-source-id: 976a4ea399ce015fc8102fb0427c4527e66bf845