ShipIt Preview Command
Summary:
This command takes a changeset as an input and prints a preview of the processed output changesets that would result.

All projects in the config registry are considered, so a single input changeset can result in multiple output changesets.

The serialization format is a little weird, but I wanted to support the ability to keep a long-running preview process using unix pipes for RPC, while also supporting the basic case of just executing one preview from the commandline.

Reviewed By: sdwilsh

Differential Revision: D6055487

fbshipit-source-id: 8554634570e16ebd492cd1f0a6d096f7d678a60d