Update source dir for github/react-vr
Summary:
React VR is moving from a git repo to the ovrsource hg monorepo. I've added ovrsource as a valid shipit repo, and updated the paths for the react-vr configuration. While this process happens, I've turned off shipit jobs for the repo (we're not shipping any code anyways).

Next steps are finishing the import from the old dir to the new one, and then testing actual commits.

Reviewed By: sdwilsh

Differential Revision: D6823773

fbshipit-source-id: f20babfe2a30305a4d8375ead2a80e3cb1c356f2