Clarify error message when importing from an unknown base branch
Summary:
The previous error message was confusing to a user who attempted to import a PR against the 'gh-pages' branch.

I'd like to add even more to the message, but we can't be sure whether the correct action is to set up a new branch config or to just merge directly on GitHub (since no FB repos have any internal place to import 'gh-pages' contents to).

Reviewed By: zpao

Differential Revision: D6970402

fbshipit-source-id: 0380cb4917ce5919518e351d2f20724b91d1192b