Add new --verify-source-commit flag
Summary:
This flag works alongside the `--verify`, `--create-fixup-patch` flags, specifying the source revision from which to verify.

This can be useful for a project that hasn't synced in an extended period of time. Rather than fixup to current HEAD, it may make sense to fixup to a specific revision, then allow the normal commit-by-commit syncing to continue forward.

Reviewed By: yns88

Differential Revision: D8043893

fbshipit-source-id: af986cf6b0597b195d6a29538f4288c3eb31c1b2