Process PR closing message before stripping SEV summary
Summary: It's not super likely that this will matter but I came across this case while testing if the PR closing message would get stripped. We still won't end up with the closing message in the diff (maybe we should try) but we at least add the number into the commit message.

Reviewed By: sdwilsh

Differential Revision: D8706564

fbshipit-source-id: e722949dac42ddf5599c4d5e0be2e6915d6d5c8b