Update ShipIt preview command to detect renames and render an error on the full patch
Summary:
The ShipIt API can't fully handle renames, because there is not enough information in the raw input patch to render a full raw output patch. We can still render the 'lite' parts of the preview, though.

This patch also includes a hack to put the error message in the 'patch' field, because the UI and Thrift API currently don't support error messages outside of a full preview crash.

Reviewed By: zpao

Differential Revision: D8842289

fbshipit-source-id: 1cb4aa0db30783dd91839a31327d801117a47d43