add git lfs support
Summary:
in  a  git or hg repo, when you install the git-lfs, and create a .gitattributes file, which include which kind of file extension you want to track with lfs, the tracked files would be sent to a designed lfs server;

so when you commit a lfs tracked file into your repo, the meta data of the file (what lfs protocol to get the file, what hash of the file) will be committed into the  git repo and the real file itself is sent to the lfs server.

in order to sync the repo from internal to external github, we need:
1. apply the  changesets( revisions) from internal to external, which is already there;
2. push the changesets (already there)
3. pull the lfs objects from internal lfs server;
4. push the lfs objects to external git server;
I have the above step **3 &4** into one phase called `ShipItPushLfsPhase`. the tricky part is explicitly specifying different lfs servers at pulling & pushing stages, which is done by specifying `lfs.url` and `lfs.pushurl`.

Reviewed By: yns88

Differential Revision: D10007601

fbshipit-source-id: bc9b790c6083eb4669cf4c3d55e4389ea1786c02