Run ShipIt in PHP7 mode
Summary:
This should move ShipIt completely over to PHP7 mode internally.

The shim for 'php' commands just adds about 300ms to the command runtime, so IMO it will be fine to leave that in place as people slowly move on to using the HHVM wrapper instead over the upcoming months.

Reviewed By: zpao

Differential Revision: D7292683

fbshipit-source-id: 326269b379da44bba82b8328184b2856ee869b9a