Configure ShipIt to sync to public osquery repo
Summary:
We were using `osquery/osquery-private` as a temporary repo during the move to Buck. The lockdown is over and the initial commits are setup, so setup ShipIt to sync to the open-source repo `facebook/osquery`.

Initially we're going to sync to the `experimental` branch while the build is not stable. On the next release when things are stable all commits from experimental should be moved to master and ShipIt reconfigured to use `master` instead.

Note for future: the commits already open-sourced on `experimental` should be directly moved to `master` (with something like `git merge experimental` for example) to avoid having new commits created by ShipIt. Once all commits are in master we can simply configure ShipIt to sync to `master` and everything should work.

Reviewed By: guliashvili

Differential Revision: D13397660

fbshipit-source-id: a25b5e2712143473482846fd4c6a5d3a744c90bf